requirements.txt
# ViralShorts V2 - Professional "Would You Rather" Video Generator
# Zero-Cost Stack for Viral YouTube Shorts

# Core video editing
moviepy==1.0.3

# Text-to-Speech (Microsoft Edge TTS - FREE, no API key needed)
edge-tts>=7.0.0
# Fallback TTS (Google TTS - more reliable but less engaging)
gTTS>=2.4.0

# Image processing (for professional graphics)
# Script handles ANTIALIAS compatibility
Pillow>=9.0.0

# Numerical processing (required for video compositing)
numpy>=1.24.0

# Audio processing
pydub>=0.25.1

# Async support
aiofiles>=23.0.0

# HTTP requests (for Pexels API, Reddit trends)
requests>=2.28.0

# Google Trends integration
pytrends>=4.9.0

# AI Content Generation (MULTI-PROVIDER - All FREE!)
groq>=0.4.0
google-generativeai>=0.3.0
google-genai>=1.0.0

# YouTube upload
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Note: You also need to install ImageMagick for TextClip to work
# Windows: Download from https://imagemagick.org/script/download.php
# Linux: sudo apt install imagemagick
# Mac: brew install imagemagick
